Split out windowing/display related methods into partial class
